requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
mutagen>=1.47.0
librosa>=0.10.1
numpy>=1.26.0
google-generativeai>=0.8.0
spotipy>=2.23.0
python-dotenv>=1.0.0
watchdog>=3.0.0
